Fix definition file generation issue
Line 1134 of drafts-definitions.js was a single concatenation of the draft definition and the dropboxMode type definition.

```
declare const draft: Draft type dropboxMode = 'add' | 'overwrite'
```

Modified the Draft.d.ts file to force a blank line between when the `build.sh` script is run (concatenating the source files including the nes for Drafts and Dropbox.